Consolidate CLI functions for vhost_user_media in vhu_media
Bug: 519646531
Test: cd base/cvd && bazel build cuttlefish/package:cvd
      # Repeat with v4l2_emulated_camera_mplane
      cvd create -media=type=v4l2_emulated_camera_splane
      # Ensure pixel format changes with each type
      v4l2-ctl -d1 --all
      # Ensure it passes with no failures
      v4l2-compliance -d1 -s
