+++ base/cvd/cuttlefish/host/commands/vhost_user_media/emulated_camera_mplane/BUILD.bazel
@@ -20,10 +20,8 @@ rust_binary(
     edition = "2024",
     deps = crate_deps([
         "clap",
-        "env_logger",
         "libc",
         "log",
-        "thiserror",
         "v4l2r",
         "vhost-user-backend",
         "virtio-media",
