+++ base/cvd/cuttlefish/host/commands/vhost_user_media/emulated_camera_mplane/Cargo.toml
@@ -5,10 +5,8 @@ edition = "2024"
 
 [dependencies]
 clap = { workspace = true }
-env_logger = { workspace = true }
 libc = { workspace = true }
 log = { workspace = true }
-thiserror = { workspace = true }
 v4l2r = { workspace = true }
 vhost-user-backend = { workspace = true }
 vhu_media = { workspace = true }
