+++ base/cvd/cuttlefish/host/commands/vhost_user_media/vhu_media/BUILD.bazel
@@ -5,9 +5,15 @@ package(default_visibility = ["//visibility:public"])
 
 rust_library(
     name = "vhu_media",
-    srcs = ["src/devices.rs", "src/lib.rs"],
+    srcs = [
+        "src/cli.rs",
+        "src/devices.rs",
+        "src/lib.rs",
+    ],
     edition = "2024",
     deps = crate_deps([
+        "clap",
+        "env_logger",
         "libc",
         "log",
         "thiserror",
